feat: location route from controller to repo
Fix: Correctif de la route du Controller -> Jdbc

+++ src/main/java/fr/diginamic/echolink/application/location/port/in/LocationGetUseCase.java
@@ -9,5 +9,5 @@ public interface LocationGetUseCase {
 
     Location getById(UUID id);
 
-    List<Location> getByGeo(float latitude, float longitude, int delta);
+    List<Location> getByGeo(double latitude, double longitude, int delta);
 }
